Move test code the unit-test module
