+++ doma-core/src/test/java/org/seasar/doma/internal/jdbc/sql/BlankLineRemovalSqlParserTest.java
@@ -1,4 +1,4 @@
-package org.seasar.doma.ut;
+package org.seasar.doma.internal.jdbc.sql;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -13,8 +13,7 @@ import java.util.function.Function;
 import org.junit.jupiter.api.Test;
 import org.seasar.doma.internal.expr.ExpressionEvaluator;
 import org.seasar.doma.internal.expr.Value;
-import org.seasar.doma.internal.jdbc.sql.NodePreparedSqlBuilder;
-import org.seasar.doma.internal.jdbc.sql.SqlParser;
+import org.seasar.doma.internal.jdbc.mock.MockConfig;
 import org.seasar.doma.jdbc.JdbcException;
 import org.seasar.doma.jdbc.PreparedSql;
 import org.seasar.doma.jdbc.SqlBuilderSettings;
