+++ settings.gradle.kts
@@ -26,4 +26,3 @@ include("integration-test-common")
 include("integration-test-java")
 include("integration-test-java-additional")
 include("integration-test-kotlin")
-include("unit-test")
