--- unit-test/.gitignore
@@ -1,12 +1,0 @@
-/bin
-/target
-/build/
-/.apt_generated/
-/.gradle/
-/.settings/
-/.project
-/.factorypath
-/.classpath
-.idea
-out
-.sdkmanrc
