--- unit-test/build.gradle.kts
@@ -1,12 +1,0 @@
-plugins {
-    `java-library`
-}
-
-dependencies {
-    implementation(project(":doma-core"))
-    testImplementation(project(":doma-mock"))
-}
-
-java {
-    toolchain.languageVersion.set(JavaLanguageVersion.of(21))
-}
